Добавил базовый класс фрагмента DefaultFragment. Наследовал класс Similar от DefaultFragment
