+++ app/src/main/java/ru/practicum/android/diploma/Vacancy.kt
@@ -21,7 +21,7 @@ class Vacancy : Fragment() {
     // TODO: Rename and change types of parameters
     private var vacancy: String? = null
 
-    private lateinit var _binding: FragmentVacancyBinding
+    private var _binding: FragmentVacancyBinding? = null
     private val binding get() = _binding!!
 
     private fun setUiListeners() {
@@ -67,7 +67,7 @@ class Vacancy : Fragment() {
          * Use this factory method to create a new instance of
          * this fragment using the provided parameters.
          *
-         * @param param1 Vacancy model
+         * @param vacancyModel Vacancy model
          * @return A new instance of fragment Vacancy.
          */
         // TODO: Rename and change types and number of parameters
